Move units to a module of its own
