+++ src/base/coordinate-grid-mixin.js
@@ -3,7 +3,8 @@ import * as d3 from 'd3';
 import {ColorMixin} from './color-mixin';
 import {MarginMixin} from './margin-mixin';
 import {BaseMixin} from './base-mixin';
-import {optionalTransition, transition, units} from '../core/core';
+import {optionalTransition, transition} from '../core/core';
+import {units} from '../core/units';
 import {constants} from '../core/constants';
 import {utils} from '../core/utils';
 import {logger} from '../core/logger';
